+++ src/common/hooks/useOrderLinesAbandonedHoldingsCheck/useOrderLinesAbandonedHoldingsCheck.test.js
@@ -4,10 +4,6 @@ import { renderHook, waitFor } from '@folio/jest-config-stripes/testing-library/
 import { useOkapiKy } from '@folio/stripes/core';
 
 import { orderLine } from 'fixtures';
-import {
-  checkIndependentPOLinesAbandonedHoldings,
-  checkSynchronizedPOLinesAbandonedHoldings,
-} from '../../utils';
 import { UNOPEN_ORDER_ABANDONED_HOLDINGS_TYPES } from '../../constants';
 import { useOrderLinesAbandonedHoldingsCheck } from './useOrderLinesAbandonedHoldingsCheck';
 
@@ -16,11 +12,6 @@ jest.mock('@folio/stripes-acq-components', () => ({
   useCentralOrderingContext: jest.fn(() => ({ isCentralOrderingEnabled: false })),
   useConsortiumTenants: jest.fn(() => ({ tenants: [] })),
   usePublishCoordinator: jest.fn(() => ({ initPublicationRequest: jest.fn() })),
-}));
-jest.mock('../../utils', () => ({
-  ...jest.requireActual('../../utils'),
-  checkIndependentPOLinesAbandonedHoldings: jest.fn(() => () => ({})),
-  checkSynchronizedPOLinesAbandonedHoldings: jest.fn(() => () => ({})),
 }));
 
 const poLines = [
@@ -45,12 +36,6 @@ const kyMock = {
 
 describe('useOrderLinesAbandonedHoldingsCheck', () => {
   beforeEach(() => {
-    checkIndependentPOLinesAbandonedHoldings
-      .mockClear()
-      .mockReturnValue(() => ({ willAbandoned: false }));
-    checkSynchronizedPOLinesAbandonedHoldings
-      .mockClear()
-      .mockReturnValue(() => ({ willAbandoned: false }));
     useOkapiKy
       .mockClear()
       .mockReturnValue(kyMock);
@@ -68,8 +53,6 @@ describe('useOrderLinesAbandonedHoldingsCheck', () => {
   });
 
   it('should return \'independent\' type in result if there is at least one abandoned holding related to independent PO Line and not to synchronized', async () => {
-    checkIndependentPOLinesAbandonedHoldings.mockReturnValue(() => ({ willAbandoned: true }));
-
     const { result } = renderHook(
       () => useOrderLinesAbandonedHoldingsCheck(poLines),
       { wrapper },
@@ -81,8 +64,6 @@ describe('useOrderLinesAbandonedHoldingsCheck', () => {
   });
 
   it('should return \'synchronized\' type in result if there is at least one abandoned holding related to synchronized PO Line', async () => {
-    checkSynchronizedPOLinesAbandonedHoldings.mockReturnValue(() => ({ willAbandoned: true }));
-
     const { result } = renderHook(
       () => useOrderLinesAbandonedHoldingsCheck(poLines),
       { wrapper },